backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
firebase-admin
moviepy
modelscope
python-dotenv
